requirements.txt
google-cloud-bigquery==2.4.0
streamlit==0.71.0
pandas==1.1.4
altair==4.1.0
